+++ Vault/src/utopia/vault/nosql/targeting/many/AccessManyRows.scala
@@ -2,14 +2,11 @@ package utopia.vault.nosql.targeting.many
 
 import utopia.flow.collection.CollectionExtensions._
 import utopia.flow.collection.immutable.Pair
-import utopia.flow.operator.enumeration.End
-import utopia.flow.operator.enumeration.End.{First, Last}
 import utopia.flow.view.immutable.caching.Lazy
 import utopia.vault.model.enumeration.SelectTarget
 import utopia.vault.model.immutable.{Column, Result, Row, Table}
 import utopia.vault.model.template.Joinable
 import utopia.vault.nosql.factory.row.FromRowFactory
-import utopia.vault.nosql.targeting.one.TargetingOne
 import utopia.vault.nosql.template.Deprecatable
 import utopia.vault.sql.{Condition, JoinType, OrderBy, SqlTarget}
 
@@ -47,7 +44,7 @@ object AccessManyRows
 	                                       f: Row => Option[A], accessCondition: Option[Condition] = None,
 	                                       ordering: Option[OrderBy] = None, limit: Option[Int] = None, offset: Int = 0,
 	                                       limitsToUniqueIndices: Boolean = false)
-		extends AccessManyRows[A]
+		extends ConcreteAccessManyLike[A, AccessManyRows[A]] with AccessManyRows[A]
 	{
 		// ATTRIBUTES   -----------------
 		
@@ -57,26 +54,16 @@ object AccessManyRows
 		// IMPLEMENTED  -----------------
 		
 		override protected def self = this
+		override protected def limitedToOne = if (limit.contains(1)) this else withLimit(1)
 		
-		override def apply(condition: Condition) = copy(accessCondition = Some(condition))
 		override def withOrdering(ordering: OrderBy) = copy(ordering = Some(ordering))
 		override def withLimit(limit: Int) = copy(limit = Some(limit))
 		override def withOffset(offset: Int, limit: Option[Int]) =
 			copy(offset = offset, limit = limit)
 		override def withLimitToUniqueIndices(limit: Boolean) = copy(limitsToUniqueIndices = limit)
 		
-		override def join(joins: Seq[Joinable], joinType: JoinType) = {
-			// WET WET (from AccessMany)
-			if (joins.isEmpty)
-				this
-			else {
-				val newTarget = joins.foldLeft(target) { _.join(_, joinType) }
-				if (newTarget == target)
-					this
-				else
-					copy(target = newTarget)
-			}
-		}
+		override protected def copyAccess(target: SqlTarget, accessCondition: Option[Condition]) =
+			copy(target = target, accessCondition = accessCondition)
 		
 		override protected def parse(row: Row) = f(row)
 		
@@ -101,32 +88,6 @@ object AccessManyRows
 					.flatMap { rows => parse(rows.head).map { _ -> rows } }
 					.toOptimizedSeq)
 			}
-		
-		override def apply(end: End, ordering: Option[OrderBy]) = {
-			val limited = if (limit.contains(1)) this else withLimit(1)
-			val ordered = {
-				if (ordering.isEmpty && (end == First || limited.ordering.isEmpty))
-					limited
-				else
-					ordering match {
-						case Some(ordering) =>
-							limited.withOrdering(end match {
-								case First => ordering
-								case Last => -ordering
-							})
-						case None =>
-							limited.ordering match {
-								case Some(ordering) =>
-									end match {
-										case First => limited
-										case Last => limited.withOrdering(-ordering)
-									}
-								case None => limited
-							}
-					}
-			}
-			TargetingOne.headOf[AccessManyRows[A], A](ordered)
-		}
 		
 		
 		// OTHER    ------------------------